3D_Multi_Modal_DL/requirements.txt
easydict==1.9
einops==0.4.1
gdown
hydra-colorlog
hydra-core
imageio==2.34.1
matplotlib==3.4.3
numpy==1.24.4
omegaconf==2.3.0
open3d==0.13.0
opencv_python==4.10.0.84
pandas==1.3.4
protobuf==3.19.1
pytorch_lightning==2.3.0
PyYAML==6.0
ruamel.base==1.0.0
ruamel_yaml==0.18.6
scikit_learn==1.0.1
scipy==1.10.1
setuptools==58.0.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tqdm==4.66.4
trimesh==4.4.1
ultralytics==8.0.135
